+++ src/main/scala/ulisse/entities/simulation/data/EngineState.scala
@@ -4,28 +4,28 @@ package ulisse.entities.simulation.data
 trait EngineState:
 
   /** The last update time */
-  def lastUpdate: Option[Double]
+  def lastUpdate: Option[Long]
 
   /** The last delta time */
-  def lastDelta: Double
+  def lastDelta: Long
 
   /** The elapsed cycle time */
-  def elapsedCycleTime: Double
+  def elapsedCycleTime: Long
 
   /** The elapsed cycle time increment */
-  def updateElapsedCycleTime(delta: Double): EngineState
+  def updateElapsedCycleTime(delta: Long): EngineState
 
   /** function to update the engine state */
-  def update(currentUpdate: Double): EngineState
+  def update(currentUpdate: Long): EngineState
 
 /** Factory for [[EngineState]] instances */
 object EngineState:
-  val defaultLastUpdate: Option[Double] = None
-  val defaultLastDelta: Double          = 0
-  val defaultElapsedCycleTime: Double   = 0
+  val defaultLastUpdate: Option[Long] = None
+  val defaultLastDelta: Long          = 0
+  val defaultElapsedCycleTime: Long   = 0
 
   /** Create a new Engine state, allowing only non negative lastUpdate and positive lastDelta and elapsedCycleTime, otherwise substitute the wrong values with defaults */
-  def apply(lastUpdate: Option[Double], lastDelta: Double, elapsedCycleTime: Double): EngineState =
+  def apply(lastUpdate: Option[Long], lastDelta: Long, elapsedCycleTime: Long): EngineState =
     EngineStateImpl(
       lastUpdate.filter(_ > 0),
       math.max(defaultLastDelta, lastDelta),
@@ -36,12 +36,12 @@ object EngineState:
   def empty(): EngineState = EngineState(defaultLastUpdate, defaultLastDelta, defaultElapsedCycleTime)
 
   private final case class EngineStateImpl(
-      lastUpdate: Option[Double],
-      lastDelta: Double,
-      elapsedCycleTime: Double
+      lastUpdate: Option[Long],
+      lastDelta: Long,
+      elapsedCycleTime: Long
   ) extends EngineState:
-    override def updateElapsedCycleTime(delta: Double): EngineState = copy(elapsedCycleTime = elapsedCycleTime + delta)
-    override def update(currentUpdate: Double): EngineState = lastUpdate match
+    override def updateElapsedCycleTime(delta: Long): EngineState = copy(elapsedCycleTime = elapsedCycleTime + delta)
+    override def update(currentUpdate: Long): EngineState = lastUpdate match
       case None => copy(lastUpdate = Some(currentUpdate))
       case Some(lastUpdate) if currentUpdate >= lastUpdate =>
         val deltaElapsed = currentUpdate - lastUpdate
